refactor: restructure HTTPDaemonClient.send() for domain-based dispatch
